Move UI-related code from "ImportPeakBoundaries" into separate class "PeakBoundaryAlerts".
